Move ranges integration tests
